+++ doc/source/extended_documentation/calibration/ensemble_calibration/create_coefficients_cube.rst
