+++ improver_tests/calibration/ensemble_calibration/test_ContinuousRankedProbabilityScoreMinimisers.py
@@ -40,10 +40,9 @@ import iris
 import numpy as np
 from iris.tests import IrisTest
 
-from improver.statistical_postprocessing.ensemble_calibration import \
+from improver.calibration.ensemble_calibration import \
     ContinuousRankedProbabilityScoreMinimisers as Plugin
-from improver.statistical_postprocessing.utilities import (
-    convert_cube_data_to_2d)
+from improver.calibration.utilities import convert_cube_data_to_2d
 from improver.utilities.warnings_handler import ManageWarnings
 
 from .helper_functions import EnsembleCalibrationAssertions, SetupCubes
