+++ improver_tests/calibration/utilities/test_utilities.py
@@ -39,7 +39,7 @@ import iris
 import numpy as np
 from iris.tests import IrisTest
 
-from improver.statistical_postprocessing.utilities import (
+from improver.calibration.utilities import (
     check_predictor_of_mean_flag, convert_cube_data_to_2d,
     flatten_ignoring_masked_data)
