+++ improver_tests/nbhood/nbhood/test_BaseNeighbourhoodProcessing.py
@@ -45,7 +45,7 @@ from improver.nbhood.circular_kernel import CircularNeighbourhood
 from improver.nbhood.nbhood import BaseNeighbourhoodProcessing as NBHood
 from improver.nbhood.nbhood import SquareNeighbourhood
 
-from ...statistical_postprocessing.ensemble_calibration.helper_functions \
+from ...calibration.ensemble_calibration.helper_functions \
     import add_forecast_reference_time_and_forecast_period
 
 SINGLE_POINT_RANGE_3_CENTROID = np.array([
